finished refactoring to use OOP TODO: refactor main.py to call the created objects
